fix sitemap rendering newlines
